third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-show-ranges-expected.txt

Tests showing several node ranges in the Summary view of detailed heap snapshot.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll
Row found: true

Step 3
Retrieved ranges: [{"from":0,"to":10}]
[0] A 77349  2  49  2  
[1] A 21348  2  48  2  
[2] A 61348  2  48  2  
[3] A 85347  2  47  2  
[4] A 181346  2  46  2  
[5] A 41345  2  45  2  
[6] A 141345  2  45  2  
[7] A 185344  2  44  2  
[8] A 57343  2  43  2  
[9] A 53341  2  41  2  
[10] Show 10 beforeShow all 40Show 10 after

Step 4
Retrieved ranges: [{"from":0,"to":10},{"from":30,"to":40}]
[0] A 77349  2  49  2  
[1] A 21348  2  48  2  
[2] A 61348  2  48  2  
[3] A 85347  2  47  2  
[4] A 181346  2  46  2  
[5] A 41345  2  45  2  
[6] A 141345  2  45  2  
[7] A 185344  2  44  2  
[8] A 57343  2  43  2  
[9] A 53341  2  41  2  
[10] Show 10 beforeShow all 20Show 10 after
[11] A 133324  1  24  1  
[12] A 17323  1  23  1  
[13] A 145323  1  23  1  
[14] A 177321  1  21  1  
[15] A 197321  1  21  1  
[16] A 189318  1  18  1  
[17] A 25316  1  16  1  
[18] A 109315  1  15  1  
[19] A 201315  1  15  1  
[20] A 193313  1  13  1  
[21] Show 10 beforeShow all 10Show 10 after

Step 5
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":30,"to":40}]
[0] A 77349  2  49  2  
[1] A 21348  2  48  2  
[2] A 61348  2  48  2  
[3] A 85347  2  47  2  
[4] A 181346  2  46  2  
[5] A 41345  2  45  2  
[6] A 141345  2  45  2  
[7] A 185344  2  44  2  
[8] A 57343  2  43  2  
[9] A 53341  2  41  2  
[10] Show 10 beforeShow all 10Show 10 after
[11] A 129334  2  34  2  
[12] A 81331  2  31  2  
[13] A 157330  2  30  2  
[14] A 137329  1  29  1  
[15] A 169329  1  29  1  
[16] Show 10 beforeShow all 5Show 10 after
[17] A 133324  1  24  1  
[18] A 17323  1  23  1  
[19] A 145323  1  23  1  
[20] A 177321  1  21  1  
[21] A 197321  1  21  1  
[22] A 189318  1  18  1  
[23] A 25316  1  16  1  
[24] A 109315  1  15  1  
[25] A 201315  1  15  1  
[26] A 193313  1  13  1  
[27] Show 10 beforeShow all 10Show 10 after

Step 6
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":28,"to":40}]
[0] A 77349  2  49  2  
[1] A 21348  2  48  2  
[2] A 61348  2  48  2  
[3] A 85347  2  47  2  
[4] A 181346  2  46  2  
[5] A 41345  2  45  2  
[6] A 141345  2  45  2  
[7] A 185344  2  44  2  
[8] A 57343  2  43  2  
[9] A 53341  2  41  2  
[10] Show 10 beforeShow all 10Show 10 after
[11] A 129334  2  34  2  
[12] A 81331  2  31  2  
[13] A 157330  2  30  2  
[14] A 137329  1  29  1  
[15] A 169329  1  29  1  
[16] Show 10 beforeShow all 3Show 10 after
[17] A 45325  1  25  1  
[18] A 33324  1  24  1  
[19] A 133324  1  24  1  
[20] A 17323  1  23  1  
[21] A 145323  1  23  1  
[22] A 177321  1  21  1  
[23] A 197321  1  21  1  
[24] A 189318  1  18  1  
[25] A 25316  1  16  1  
[26] A 109315  1  15  1  
[27] A 201315  1  15  1  
[28] A 193313  1  13  1  
[29] Show 10 beforeShow all 10Show 10 after

Step 7
Retrieved ranges: [{"from":0,"to":10},{"from":18,"to":26},{"from":28,"to":40}]
[0] A 77349  2  49  2  
[1] A 21348  2  48  2  
[2] A 61348  2  48  2  
[3] A 85347  2  47  2  
[4] A 181346  2  46  2  
[5] A 41345  2  45  2  
[6] A 141345  2  45  2  
[7] A 185344  2  44  2  
[8] A 57343  2  43  2  
[9] A 53341  2  41  2  
[10] Show 10 beforeShow all 8Show 10 after
[11] A 125335  2  35  2  
[12] A 117334  2  34  2  
[13] A 129334  2  34  2  
[14] A 81331  2  31  2  
[15] A 157330  2  30  2  
[16] A 137329  1  29  1  
[17] A 169329  1  29  1  
[18] A 205328  1  28  1  
[19] Show 10 beforeShow all 2Show 10 after
[20] A 45325  1  25  1  
[21] A 33324  1  24  1  
[22] A 133324  1  24  1  
[23] A 17323  1  23  1  
[24] A 145323  1  23  1  
[25] A 177321  1  21  1  
[26] A 197321  1  21  1  
[27] A 189318  1  18  1  
[28] A 25316  1  16  1  
[29] A 109315  1  15  1  
[30] A 201315  1  15  1  
[31] A 193313  1  13  1  
[32] Show 10 beforeShow all 10Show 10 after

Step 8
Retrieved ranges: [{"from":0,"to":10},{"from":15,"to":45}]
[0] A 77349  2  49  2  
[1] A 21348  2  48  2  
[2] A 61348  2  48  2  
[3] A 85347  2  47  2  
[4] A 181346  2  46  2  
[5] A 41345  2  45  2  
[6] A 141345  2  45  2  
[7] A 185344  2  44  2  
[8] A 57343  2  43  2  
[9] A 53341  2  41  2  
[10] Show 10 beforeShow all 5Show 10 after
[11] A 65338  2  38  2  
[12] A 113338  2  38  2  
[13] A 153336  2  36  2  
[14] A 125335  2  35  2  
[15] A 117334  2  34  2  
[16] A 129334  2  34  2  
[17] A 81331  2  31  2  
[18] A 157330  2  30  2  
[19] A 137329  1  29  1  
[20] A 169329  1  29  1  
[21] A 205328  1  28  1  
[22] A 37327  1  27  1  
[23] A 49326  1  26  1  
[24] A 45325  1  25  1  
[25] A 33324  1  24  1  
[26] A 133324  1  24  1  
[27] A 17323  1  23  1  
[28] A 145323  1  23  1  
[29] A 177321  1  21  1  
[30] A 197321  1  21  1  
[31] A 189318  1  18  1  
[32] A 25316  1  16  1  
[33] A 109315  1  15  1  
[34] A 201315  1  15  1  
[35] A 193313  1  13  1  
[36] A 149312  1  12  1  
[37] A 69311  1  11  1  
[38] A 2939  0  9  0  
[39] A 12138  0  8  0  
[40] A 16538  0  8  0  
[41] Show 10 beforeShow all 5Show 10 after

Step 9
Retrieved ranges: [{"from":0,"to":45}]
[0] A 77349  2  49  2  
[1] A 21348  2  48  2  
[2] A 61348  2  48  2  
[3] A 85347  2  47  2  
[4] A 181346  2  46  2  
[5] A 41345  2  45  2  
[6] A 141345  2  45  2  
[7] A 185344  2  44  2  
[8] A 57343  2  43  2  
[9] A 53341  2  41  2  
[10] A 89340  2  40  2  
[11] A 97340  2  40  2  
[12] A 73339  2  39  2  
[13] A 101339  2  39  2  
[14] A 13338  2  38  2  
[15] A 65338  2  38  2  
[16] A 113338  2  38  2  
[17] A 153336  2  36  2  
[18] A 125335  2  35  2  
[19] A 117334  2  34  2  
[20] A 129334  2  34  2  
[21] A 81331  2  31  2  
[22] A 157330  2  30  2  
[23] A 137329  1  29  1  
[24] A 169329  1  29  1  
[25] A 205328  1  28  1  
[26] A 37327  1  27  1  
[27] A 49326  1  26  1  
[28] A 45325  1  25  1  
[29] A 33324  1  24  1  
[30] A 133324  1  24  1  
[31] A 17323  1  23  1  
[32] A 145323  1  23  1  
[33] A 177321  1  21  1  
[34] A 197321  1  21  1  
[35] A 189318  1  18  1  
[36] A 25316  1  16  1  
[37] A 109315  1  15  1  
[38] A 201315  1  15  1  
[39] A 193313  1  13  1  
[40] A 149312  1  12  1  
[41] A 69311  1  11  1  
[42] A 2939  0  9  0  
[43] A 12138  0  8  0  
[44] A 16538  0  8  0  
[45] Show 10 beforeShow all 5Show 10 after

Profiler was disabled.
